adjust resource costs from items when building from attacks

+++ app/Domain/Interfaces/AdjustsResourceCostAmounts.php
@@ -6,7 +6,7 @@ namespace App\Domain\Interfaces;
 
 interface AdjustsResourceCostAmounts
 {
-    public function adjustResourceCostAmount(float $amount): float;
+    public function adjustResourceCostAmount(float $amount): int;
 
     public function adjustResourceCostPercent(float $amount): float;
 }
